requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.1.1
open_clip_torch==2.26.1
openai==1.12.0
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
requests==2.32.3
numpy==1.26.4
opencv-python-headless==4.10.0.84
psycopg[binary]==3.2.1
pgvector==0.3.2

